Refactor task handlers to be UI specific
